\name{ebma-package}
\alias{ebma-package}
\alias{ebma}
\docType{package}
\title{
  Ensemble Bayesian Model Averaging (EBMA)
}
\description{
  I honestly don't understand a lot of this package so couldn't tell ya unfortunately 
}
\details{
  EBMA package with functions to calculate the z hat values, weights, and the overall EBMA vector. 
}
\author{
Alex Avery

Maintainer: Alex Avery <a.d.avery@wustl.edu>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
